Move groups into admin and lift url helpers into visualizations
