+++ enterprise/frontend/src/metabase-enterprise/audit_app/metabot-analytics/components/ConversationDetailPage/ConversationHeader.tsx
@@ -2,6 +2,7 @@ import { useMemo } from "react";
 import { msgid, ngettext, t } from "ttag";
 
 import { getGroupFocusPermissionsUrl } from "metabase/admin/permissions/utils/urls";
+import { isAdminGroup, isDefaultGroup } from "metabase/admin/utils/groups";
 import {
   skipToken,
   useListPermissionsGroupsQuery,
@@ -23,7 +24,6 @@ import {
   Title,
 } from "metabase/ui";
 import * as Urls from "metabase/urls";
-import { isAdminGroup, isDefaultGroup } from "metabase/utils/groups";
 import { getUserName } from "metabase/utils/user";
 import { useGetTenantQuery } from "metabase-enterprise/api";
 import * as EnterpriseUrls from "metabase-enterprise/urls";
