+++ frontend/src/metabase/admin/people/components/GroupMembersTable/GroupMembersTable.tsx
@@ -3,6 +3,7 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import { AdminContentTable } from "metabase/admin/components/AdminContentTable";
+import { isAdminGroup, isDefaultGroup } from "metabase/admin/utils/groups";
 import { Link } from "metabase/common/components/Link";
 import { PaginationControls } from "metabase/common/components/PaginationControls";
 import { usePagination } from "metabase/common/hooks/use-pagination";
@@ -10,7 +11,6 @@ import { PLUGIN_GROUP_MANAGERS, PLUGIN_TENANTS } from "metabase/plugins";
 import { useSelector } from "metabase/redux";
 import { getUser } from "metabase/selectors/user";
 import { Box, Flex, Icon, Text, Tooltip, UnstyledButton } from "metabase/ui";
-import { isAdminGroup, isDefaultGroup } from "metabase/utils/groups";
 import { getFullName } from "metabase/utils/user";
 import type { Group, Member, Membership } from "metabase-types/api";
