+++ frontend/src/metabase/admin/people/components/GroupSummary.tsx
@@ -1,7 +1,7 @@
 import { msgid, ngettext, t } from "ttag";
 
+import { isAdminGroup, isDefaultGroup } from "metabase/admin/utils/groups";
 import { Box, type BoxProps } from "metabase/ui";
-import { isAdminGroup, isDefaultGroup } from "metabase/utils/groups";
 import type { GroupInfo } from "metabase-types/api";
 
 interface GroupSummaryProps extends BoxProps {
