+++ frontend/src/metabase/admin/people/components/MembershipSelect/MembershipSelect.tsx
@@ -2,14 +2,14 @@ import { useDisclosure } from "@mantine/hooks";
 import { useMemo } from "react";
 import { t } from "ttag";
 
-import { PLUGIN_GROUP_MANAGERS, PLUGIN_TENANTS } from "metabase/plugins";
-import { Box, Divider, Flex, Icon, Popover } from "metabase/ui";
 import {
   getGroupNameLocalized,
   isAdminGroup,
   isDataAnalystGroup,
   isDefaultGroup,
-} from "metabase/utils/groups";
+} from "metabase/admin/utils/groups";
+import { PLUGIN_GROUP_MANAGERS, PLUGIN_TENANTS } from "metabase/plugins";
+import { Box, Divider, Flex, Icon, Popover } from "metabase/ui";
 import { isNotNull } from "metabase/utils/types";
 import type { GroupInfo, Member } from "metabase-types/api";
