+++ frontend/src/metabase/admin/people/forms/UserForm.tsx
@@ -5,6 +5,7 @@ import _ from "underscore";
 import * as Yup from "yup";
 
 import { MembershipSelect } from "metabase/admin/people/components/MembershipSelect";
+import { isAdminGroup, isDefaultGroup } from "metabase/admin/utils/groups";
 import { useListPermissionsGroupsQuery } from "metabase/api";
 import { FormField } from "metabase/common/components/FormField";
 import { FormFooter } from "metabase/common/components/FormFooter";
@@ -21,7 +22,6 @@ import {
 } from "metabase/plugins";
 import { Button } from "metabase/ui";
 import * as Errors from "metabase/utils/errors";
-import { isAdminGroup, isDefaultGroup } from "metabase/utils/groups";
 import type { GroupId, Member, User, UserId } from "metabase-types/api";
 
 const localUserSchema = Yup.object({
