+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/groups.ts
@@ -1,15 +1,15 @@
 import { createSelector } from "@reduxjs/toolkit";
 import _ from "underscore";
 
-import { permissionApi } from "metabase/api";
-import { PLUGIN_TENANTS } from "metabase/plugins";
-import type { State } from "metabase/redux/store";
 import {
   getGroupSortOrder,
   isAdminGroup,
   isDataAnalystGroup,
   isDefaultGroup,
-} from "metabase/utils/groups";
+} from "metabase/admin/utils/groups";
+import { permissionApi } from "metabase/api";
+import { PLUGIN_TENANTS } from "metabase/plugins";
+import type { State } from "metabase/redux/store";
 import type { Group, GroupId } from "metabase-types/api";
 
 const isPinnedGroup = (group: Group) =>
