+++ frontend/src/metabase/dashboard/visualizations/LinkViz/LinkViz.tsx
@@ -8,7 +8,7 @@ import { getParameterValues } from "metabase/dashboard/selectors";
 import { SearchResults } from "metabase/nav/components/search/SearchResults";
 import { useSelector } from "metabase/redux";
 import { modelToUrl } from "metabase/urls";
-import { getUrlTarget } from "metabase/utils/dom";
+import { getUrlTarget } from "metabase/visualizations/lib/open-url";
 import { fillParametersInText } from "metabase/visualizations/shared/utils/parameter-substitution";
 import type {
   Dashboard,
