+++ frontend/src/metabase/query_builder/actions/core/core.ts
@@ -24,9 +24,9 @@ import type { Dispatch, GetState } from "metabase/redux/store";
 import { getMetadata } from "metabase/selectors/metadata";
 import * as Urls from "metabase/urls";
 import { clone } from "metabase/utils/clone";
-import { shouldOpenInBlankWindow } from "metabase/utils/dom";
 import { isNotNull } from "metabase/utils/types";
 import { getDefaultSize } from "metabase/visualizations";
+import { shouldOpenInBlankWindow } from "metabase/visualizations/lib/open-url";
 import { getCardAfterVisualizationClick } from "metabase/visualizations/lib/utils";
 import * as Lib from "metabase-lib";
 import Question from "metabase-lib/v1/Question";
