+++ frontend/src/metabase/visualizations/lib/action.ts
@@ -3,7 +3,6 @@ import { push } from "react-router-redux";
 import _ from "underscore";
 
 import type { Dispatch } from "metabase/redux/store";
-import { open } from "metabase/utils/dom";
 import type Question from "metabase-lib/v1/Question";
 
 import type {
@@ -11,6 +10,8 @@ import type {
   OnChangeCardAndRun,
   QuestionChangeClickAction,
 } from "../types";
+
+import { openUrl } from "./open-url";
 
 type ActionProps = {
   dispatch: Dispatch;
@@ -43,7 +44,7 @@ export function performAction(
     const url = action.url();
     const ignoreSiteUrl = action.ignoreSiteUrl;
     if (url) {
-      open(url, {
+      openUrl(url, {
         openInSameOrigin: (location) => {
           if (props.onSameOriginNavigation) {
             props.onSameOriginNavigation(location);
